attached_assets/flask_kpn/kpn/requirements.txt
flask>=3.1.2
flask-caching>=2.3.1
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.2
werkzeug>=3.1.3
pillow>=11.3.0
redis>=6.4.0
facebook-sdk
SQLAlchemy
Jinja2
requests>=2.32.5
requests-oauthlib>=2.0.0
cryptography>=46.0.1
psycopg2-binary>=2.9.10
gunicorn>=21.2.0
python-dotenv>=1.0.1